backend/pyproject.toml
[project]
name = "codesentinel-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "arq>=0.25.0",
    "redis>=5.0.0",
    "httpx>=0.26.0",
    "structlog>=24.1.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "anyio[trio]>=4.2.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.26.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["alembic*", "tests*", "dist*", "*.egg-info"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --ignore=tests/test_live.py"  # exclude live tests from default run
markers = [
    "live: requires full Docker stack running (deselected by default)",
]

[tool.ruff]
line-length = 100
target-version = "py311"
